refactor: move checkFim and FimResult to utils for easier testing
